[#109][#187] Lazify interpretters, create schemata generator script (#197)
* [#193][#192] Add WithMap, WithOption, WithOptional, OptionFromUndefined, OptionFromNullable

* [#193] correct WithOption > Type, add jsdoc to WithDate

* [#166] move modules into base

* [#109] unify schemable schemas

* [#187] lazify interpretters, add type export to main files

* [#109] add schemata generator script

* [#109] make prettier sync

* chore: ignore schemata coverage

* chore: add docs and docs to schemata script

* chore: fix hsl example

* [#187] de-lazify interpretters

+++ tests/schemables/WithFloat.test.ts
@@ -81,7 +81,7 @@ describe('NegativeFloat', () => {
   })
 
   describe('Schema', () => {
-    const Float = WithFloat.Schema.float()
+    const Float = WithFloat.Schema()
     it('derives a decoder', () => {
       const decoder = getDecoder(Float)
       expect(decoder.decode(NaN)._tag).toEqual('Left')
